16:9 aspect ratio for front camera instead of square + new reward function
More details: Episodes now stop automatically when the vehicle stalls for too long. The ego vehicle and the NPC vehicles only spawn on the highway. A new reward function was implemented to avoid learning to drive in circles. The outputs of the camera/inputs to the encoder of the model can now be rectangular, allowing for more suitable wide camera shots for autonomous driving.